pyproject.toml
[tool.poetry]
name = "tank-viewer"
version = "0.1.0"
description = "Visualizer for seed tank data from a chain of devices via an Arduino (UART)"
authors = ["Jeremy Rempel <jeremy.rempel@usask.ca>"]

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "^3.5.1"
pyserial = "^3.5"
cycler = "^0.11.0"
fonttools = "^4.29.1"
kiwisolver = "^1.3.2"
numpy = "^1.22.2"
packaging = "^21.3"
Pillow = "^9.0.1"
pyparsing = "^3.0.7"
python-dateutil = "^2.8.2"
setuptools-scm = "^6.4.2"
six = "^1.16.0"
tomli = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
